+++ src/api/components/settings/settings.controller.js
@@ -84,10 +84,11 @@ exports.patchTarget = async (req, res) => {
 
       for (const camera of cameraSettings) {
         const controller = cameras.get(camera.name);
-        controller?.videoanalysis.changeZone(
+        controller?.videoanalysis.changeSettings(
           camera.videoanalysis.regions,
           camera.videoanalysis.sensitivity,
-          camera.videoanalysis.difference
+          camera.videoanalysis.difference,
+          camera.videoanalysis.dwellTimer
         );
       }
     }